started to refactor bulk create users from csv, fixing front end error display
