+++ hat/assets/js/apps/Iaso/domains/users/components/BulkImportDialog/BulkImportDialog.tsx
@@ -22,7 +22,6 @@ import { BulkImportDefaults } from '../../types';
 import { BulkImportButton } from './BulkImportButton';
 import { DefaultValuesSection } from './DefaultValuesSection';
 import { useBulkUserValidation } from './hooks/useBulkUserValidation';
-import { useErrorProcessor } from './hooks/useErrorProcessor';
 import { useUploadCsv } from './hooks/useUploadCsv';
 import { ValidationErrorTable } from './ValidationErrorTable';
 
@@ -48,7 +47,7 @@ export const BulkImportDialogModal: FunctionComponent<Props> = ({
     const [showDefaults, setShowDefaults] = useState(false);
     const [defaults, setDefaults] = useState<BulkImportDefaults>({});
 
-    const { mutateAsync: upload, isLoading } = useUploadCsv();
+    const { mutateAsync: upload, isLoading, error: error } = useUploadCsv();
 
     const {
         apiErrors,
@@ -85,13 +84,8 @@ export const BulkImportDialogModal: FunctionComponent<Props> = ({
         isValid,
     } = formik;
 
-    const { csvValidationErrors, simpleErrors } = useErrorProcessor(
-        apiErrors,
-        errors.file,
-    );
-
     const allowConfirm = isValid && Boolean(touched.file) && !isLoading;
-    const Buttons = ({ closeDialog: close }) => {
+    const Buttons = ({ closeDialog: close }: { closeDialog: () => void }) => {
         return (
             <FileUploadButtons
                 allowConfirm={allowConfirm}
@@ -105,10 +99,15 @@ export const BulkImportDialogModal: FunctionComponent<Props> = ({
         );
     };
 
+    // as formik error parsing flats everything as a string, we need to access this in a "raw" way
+    const fileContentErrors = React.useMemo(() => {
+        return error?.details?.file_content;
+    }, [error?.details]);
+
     return (
         <SimpleModal
             titleMessage={titleMessage}
-            maxWidth={csvValidationErrors.length > 0 ? 'lg' : 'sm'}
+            maxWidth={fileContentErrors?.length > 0 ? 'lg' : 'sm'}
             open={isOpen}
             closeDialog={closeDialog}
             id={id ?? 'bulk-user-create'}
@@ -132,7 +131,7 @@ export const BulkImportDialogModal: FunctionComponent<Props> = ({
                     }}
                     required
                     multi={false}
-                    errors={simpleErrors}
+                    errors={errors?.file ? [errors?.file] : undefined}
                     placeholder={formatMessage(MESSAGES.selectCsvFile)}
                 />
             </Box>
@@ -159,7 +158,7 @@ export const BulkImportDialogModal: FunctionComponent<Props> = ({
                 </Box>
             </Collapse>
 
-            {csvValidationErrors.length > 0 && (
+            {fileContentErrors?.length > 0 && (
                 <Box mt={2}>
                     <Alert severity="error" sx={{ mb: 2 }}>
                         <Typography variant="subtitle1">
@@ -167,11 +166,11 @@ export const BulkImportDialogModal: FunctionComponent<Props> = ({
                         </Typography>
                         <Typography variant="body2">
                             {formatMessage(MESSAGES.fixErrorsAndRetry, {
-                                count: csvValidationErrors.length,
+                                count: fileContentErrors.length,
                             })}
                         </Typography>
                     </Alert>
-                    <ValidationErrorTable errors={csvValidationErrors} />
+                    <ValidationErrorTable errors={fileContentErrors} />
                 </Box>
             )}
