+++ iaso/api/profiles/views.py
@@ -27,10 +27,10 @@ from rest_framework.response import Response
 
 from hat.api.export_utils import Echo, generate_xlsx, iter_items
 from hat.audit.models import PROFILE_API
+from iaso.api.bulk_create_users.constants import BULK_CREATE_USER_COLUMNS_LIST
 from iaso.api.common import CONTENT_TYPE_CSV, CONTENT_TYPE_XLSX, FileFormatEnum, ModelViewSet
 from iaso.api.filters import CamelCaseDjangoFilterBackend, CamelCaseOrderingFilter
 from iaso.api.profiles.audit import ProfileAuditLogger
-from iaso.api.profiles.bulk_create_users import BULK_CREATE_USER_COLUMNS_LIST
 from iaso.api.profiles.constants import PK_ME
 from iaso.api.profiles.filters import ProfileListFilter
 from iaso.api.profiles.pagination import ProfileDropdownPagination, ProfilePagination
