+++ iaso/tests/fixtures/test_user_bulk_create_valid_with_multiple_teams.csv
@@ -1,2 +1,2 @@
 username,password,email,first_name,last_name,orgunit,orgunit__source_ref,profile_language,dhis2_id,organization,permissions,user_roles,projects,teams,phone_number,editable_org_unit_types
-rsfg,frdghb30cd!hd,efsfg@srgrs.com,feesf,dsfsf,"Tatooine, Dagobah",,FR,dhis2_id_3,,,,,"{{team_names}}",,{{org_unit_type_id}}
+rsfg,frdghb30cd!hd,efsfg@srgrs.com,feesf,dsfsf,Tatooine; Dagobah,,FR,dhis2_id_3,,,,,"{{team_names}}",,{{org_unit_type_id}}
