+++ requirements.txt
@@ -68,6 +68,7 @@ unidecode==1.3.7  # https://github.com/avian2/unidecode/tags
 python-dateutil==2.9.0 # https://github.com/dateutil/dateutil
 names-generator==0.2.0 # https://github.com/glentner/names_generator
 django-autoslug==1.9.9 # https://github.com/justinmayer/django-autoslug
+python-magic==0.4.27 # https://github.com/ahupp/python-magic
 
 # Exporting tools.
 # ------------------------------------------------------------------------------
